+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `colorpicker` builder (https://github.com/edvin/tornadofx/issues/76)
+- `chooseFile` File Chooser (https://github.com/edvin/tornadofx/issues/76)
+- `pagination` builder (https://github.com/edvin/tornadofx/issues/76)
+- Configurable alert dialog 
 - `Node.bindClass` applies an observable style class to a node (https://github.com/edvin/tornadofx/issues/88)
 
 ### Changed